finish tle-raw topk
