--- python/tutorials/tle/raw/output.txt
@@ -1,66 +1,0 @@
-0  selected/all: 2048 / 2048 = 1.0
-1  selected/all: 2048 / 2048 = 1.0
-2  selected/all: 2048 / 2048 = 1.0
-3  selected/all: 2048 / 2048 = 1.0
-4  selected/all: 2048 / 2048 = 1.0
-5  selected/all: 2048 / 2048 = 1.0
-6  selected/all: 2048 / 2048 = 1.0
-7  selected/all: 2048 / 2048 = 1.0
-8  selected/all: 2048 / 2048 = 1.0
-9  selected/all: 2048 / 2048 = 1.0
-10  selected/all: 2048 / 2048 = 1.0
-11  selected/all: 2048 / 2048 = 1.0
-12  selected/all: 2048 / 2048 = 1.0
-13  selected/all: 2048 / 2048 = 1.0
-14  selected/all: 2048 / 2048 = 1.0
-15  selected/all: 2048 / 2048 = 1.0
-16  selected/all: 2048 / 2048 = 1.0
-17  selected/all: 2048 / 2048 = 1.0
-18  selected/all: 2048 / 2048 = 1.0
-19  selected/all: 2048 / 2048 = 1.0
-20  selected/all: 2048 / 2048 = 1.0
-21  selected/all: 2048 / 2048 = 1.0
-22  selected/all: 2048 / 2048 = 1.0
-23  selected/all: 2048 / 2048 = 1.0
-24  selected/all: 2048 / 2048 = 1.0
-25  selected/all: 2048 / 2048 = 1.0
-26  selected/all: 2048 / 2048 = 1.0
-27  selected/all: 2048 / 2048 = 1.0
-28  selected/all: 2048 / 2048 = 1.0
-29  selected/all: 2048 / 2048 = 1.0
-30  selected/all: 2048 / 2048 = 1.0
-31  selected/all: 2048 / 2048 = 1.0
-32  selected/all: 2048 / 2048 = 1.0
-33  selected/all: 2048 / 2048 = 1.0
-34  selected/all: 2048 / 2048 = 1.0
-35  selected/all: 2048 / 2048 = 1.0
-36  selected/all: 2048 / 2048 = 1.0
-37  selected/all: 2048 / 2048 = 1.0
-38  selected/all: 2048 / 2048 = 1.0
-39  selected/all: 2048 / 2048 = 1.0
-40  selected/all: 2048 / 2048 = 1.0
-41  selected/all: 2048 / 2048 = 1.0
-42  selected/all: 2048 / 2048 = 1.0
-43  selected/all: 2048 / 2048 = 1.0
-44  selected/all: 2048 / 2048 = 1.0
-45  selected/all: 2048 / 2048 = 1.0
-46  selected/all: 2048 / 2048 = 1.0
-47  selected/all: 2048 / 2048 = 1.0
-48  selected/all: 2048 / 2048 = 1.0
-49  selected/all: 2048 / 2048 = 1.0
-50  selected/all: 2048 / 2048 = 1.0
-51  selected/all: 2048 / 2048 = 1.0
-52  selected/all: 2048 / 2048 = 1.0
-53  selected/all: 2048 / 2048 = 1.0
-54  selected/all: 2048 / 2048 = 1.0
-55  selected/all: 2048 / 2048 = 1.0
-56  selected/all: 2048 / 2048 = 1.0
-57  selected/all: 2048 / 2048 = 1.0
-58  selected/all: 2048 / 2048 = 1.0
-59  selected/all: 2048 / 2048 = 1.0
-60  selected/all: 2048 / 2048 = 1.0
-61  selected/all: 2048 / 2048 = 1.0
-62  selected/all: 2048 / 2048 = 1.0
-63  selected/all: 2048 / 2048 = 1.0
-Average bucket_sort_topk time: 0.051 ms
-Average torch.topk time: 0.103 ms
